requirements-prod.txt
# 核心依赖
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
gunicorn>=21.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
alembic>=1.12.0

# 安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 工具
pyautogen>=0.2.0
python-dotenv>=1.0.0
openai>=1.0.0
duckduckgo-search>=3.9.0 